+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add configurable number of shown months
 - #51 Add unit tests with ci: [testling](https://ci.testling.com/rikkert/pikaday)
 - Fix #94 blur input field after date selection
+- Add configuable container to render calendar into 
+- Add option to show week numbers, see #147
 
 ## 1.2.0 - 2014-01-28